Encapsulate Firebase storage with images behaviour
Adds FirebaseImageManager class with download and upload methods to
handle Firebase interaction nicely
